ai-agents/requirements.txt
# CrewAI core
crewai==0.63.6
crewai-tools==0.12.0

# LangChain + Ollama
langchain==0.2.16
langchain-community==0.2.16
langchain-ollama==0.1.3

# FastAPI server
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

# OpenSearch client
opensearch-py==2.6.0

# HTTP
requests==2.32.3
urllib3==2.2.3

# Data
pyyaml==6.0.2
python-dotenv==1.0.1

# Async
anyio==4.4.0
asyncio==3.4.3
